refactor: :art: migrated account component to TSX
